reorderings and api improvements
