+++ packages/www/components/Dashboard/Table/filters/fields/new.tsx
@@ -1,0 +1,3 @@
+// TODO: Pass this to ./index.tsx when fully implemented
+
+export type FilterType = "text" | "number" | "boolean" | "date";
