+++ packages/www/components/Dashboard/Table/filters/helpers.tsx
@@ -15,7 +15,6 @@ const StyledHeader = styled(Accordion.Header, {
   display: "flex",
   borderTop: "1px solid",
   borderColor: "$slate6",
-  padding: "11px",
   alignItems: "center",
 });
 
@@ -24,7 +23,8 @@ const StyledButton = styled(Accordion.Button, {
   border: "none",
   display: "flex",
   alignItems: "center",
-  padding: 0,
+  padding: "11px",
+  width: "100%",
 });
 
 const StyledPanel = styled(Accordion.Panel, {
@@ -45,8 +45,8 @@ const FilterIcon = () => {
       fill="none"
       xmlns="http://www.w3.org/2000/svg">
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M10.4955 0.00305184C10.9118 0.00305184 11.25 0.339052 11.25 0.753052C11.2499 0.919997 11.1939 1.08211 11.091 1.21355L6.75451 6.75605L6.75002 10.2946C6.75009 10.3932 6.73072 10.4908 6.69302 10.5819C6.65531 10.673 6.6 10.7558 6.53026 10.8256L5.37826 11.9776C5.36782 11.9881 5.35449 11.9953 5.33995 11.9982C5.32541 12.0012 5.31033 11.9998 5.2966 11.9941C5.28287 11.9885 5.27112 11.979 5.26283 11.9667C5.25454 11.9544 5.25008 11.9399 5.25002 11.9251V6.75605L0.909015 1.21355C0.848148 1.13587 0.803275 1.04691 0.776989 0.951789C0.750702 0.85667 0.743523 0.757286 0.755866 0.659377C0.76821 0.561468 0.79983 0.466975 0.848902 0.381357C0.897974 0.295738 0.963524 0.220692 1.04177 0.160552C1.17437 0.0583577 1.3371 0.00297272 1.50452 0.00305184H10.4955Z"
         fill={resolvedTheme === "dark" ? "white" : "black"}
       />
@@ -99,8 +99,8 @@ const NextIcon = () => {
       fill="none"
       xmlns="http://www.w3.org/2000/svg">
       <path
-        fill-rule="evenodd"
-        clip-rule="evenodd"
+        fillRule="evenodd"
+        clipRule="evenodd"
         d="M2.85 7.2345C2.0812 7.19437 1.35709 6.86087 0.826857 6.30274C0.296627 5.7446 0.000681416 5.00434 0 4.2345C0 3.018 0.7245 1.9695 1.7655 1.5L3 2.7345C2.60217 2.7345 2.22064 2.89254 1.93934 3.17384C1.65804 3.45514 1.5 3.83668 1.5 4.2345C1.5 4.63232 1.65804 5.01386 1.93934 5.29516C2.22064 5.57646 2.60217 5.7345 3 5.7345H8.64L6.21975 3.26475L7.28025 2.20425L11.5605 6.4845L7.28025 10.7647L6.21975 9.70425L8.64 7.2345H2.85Z"
         fill={resolvedTheme === "dark" ? "white" : "black"}
       />
@@ -119,8 +119,8 @@ const CalendarIcon = () => {
       xmlns="http://www.w3.org/2000/svg">
       <g opacity="0.6">
         <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
+          fillRule="evenodd"
+          clipRule="evenodd"
           d="M0 6H12V10.5C12 10.8978 11.842 11.2794 11.5607 11.5607C11.2794 11.842 10.8978 12 10.5 12H1.5C1.10218 12 0.720644 11.842 0.43934 11.5607C0.158035 11.2794 0 10.8978 0 10.5V6ZM9.75 1.5H10.5C10.8978 1.5 11.2794 1.65804 11.5607 1.93934C11.842 2.22064 12 2.60218 12 3V4.5H0V3C0 2.60218 0.158035 2.22064 0.43934 1.93934C0.720644 1.65804 1.10218 1.5 1.5 1.5H2.25V2.25C2.25 2.44891 2.32902 2.63968 2.46967 2.78033C2.61032 2.92098 2.80109 3 3 3C3.19891 3 3.38968 2.92098 3.53033 2.78033C3.67098 2.63968 3.75 2.44891 3.75 2.25V1.5H8.25V2.25C8.25 2.44891 8.32902 2.63968 8.46967 2.78033C8.61032 2.92098 8.80109 3 9 3C9.19891 3 9.38968 2.92098 9.53033 2.78033C9.67098 2.63968 9.75 2.44891 9.75 2.25V1.5ZM3 0C3.19891 0 3.38968 0.0790176 3.53033 0.21967C3.67098 0.360322 3.75 0.551088 3.75 0.75V1.5H2.25V0.75C2.25 0.551088 2.32902 0.360322 2.46967 0.21967C2.61032 0.0790176 2.80109 0 3 0V0ZM9 0C9.19891 0 9.38968 0.0790176 9.53033 0.21967C9.67098 0.360322 9.75 0.551088 9.75 0.75V1.5H8.25V0.75C8.25 0.551088 8.32902 0.360322 8.46967 0.21967C8.61032 0.0790176 8.80109 0 9 0V0Z"
           fill={resolvedTheme === "dark" ? "white" : "black"}
         />
